+++ CHANGELOG.md
@@ -7,6 +7,8 @@ zf-queue
 - Enh: Drop `Bupy7\Queue\Service\TaskService` and
 moved `Bupy7\Queue\Service\TaskService::add()` method
 to `Bupy7\Queue\Service\QueueService::add()`.
+- Fix: Small bugs.
+- Enh: Tests.
 
 1.0.0-alpha
 -----------